Fix : vm make fail 	modified:   include/vm/uninit.h 	modified:   include/vm/vm.h 	new file:   include/vm/vm_type.h
